apps/backend/pyproject.toml
[project]
name = "backend"
version = "0.0.0"
description = "FastAPI backend application"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.6",
    "pydantic-settings>=2.7.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "aiosqlite>=0.21.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "email-validator>=2.2.0",
    "bcrypt>=4.2.0",
    "python-jose[cryptography]>=3.3.0",
    "strawberry-graphql[fastapi]>=0.262.0",
    "aioboto3>=14.0.0",
    "python-multipart>=0.0.18",
]

[dependency-groups]
dev = [
    "ruff>=0.8.4",
]
test = [
    "httpx>=0.28.1",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
    "pytest-cov>=6.0.0",
    "moto[s3]>=5.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
source = ["src"]
branch = true
concurrency = ["greenlet", "thread"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
]

[tool.ruff]
target-version = "py313"
line-length = 88
src = ["src"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
